requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
Authlib==1.3.0
requests==2.31.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
Pillow==10.1.0
numpy==1.26.0